latestapp

+++ blueprints/pembeli/resources.py
@@ -1,4 +1,4 @@
-import json, logging
+import json, logging, hashlib
 from flask import Blueprint
 from flask_restful import Api, Resource, reqparse, marshal
 from . import *
@@ -16,43 +16,54 @@ class PembeliResource(Resource):
         parse.add_argument('username', location='args', required=True)
         parse.add_argument('password', location='args', required=True)
         parse.add_argument('contact', location='args',  required=True)
-        parse.add_argument('status', location='args', default='pembeli')
         parse.add_argument('email', location='args', required=True)
         parse.add_argument('address', location='args', required=True)
 
         args = parse.parse_args()
 
-        pembeliBaru = Pembeli(None, args['username'], args['password'], args['contact'], args['status'], args['email'], args['address'])
-        db.session.add(pembeliBaru)
-        db.session.commit()        
+        qry = Pembeli.query.filter_by(username = args['username']).first()
+        if qry is not None:
+            return {'status':'Not Acceptable', 'message': 'USERNAME_ALREADY_IN_USE'}, 406, {'Content-Type': 'application/json'}
 
-        return marshal(pembeliBaru, Pembeli.response_field), 200, {'Content-Type': 'application/json'}
+        password = hashlib.md5(args['password'].encode()).hexdigest()
+
+        pembeliBaru = Pembeli(None, args['username'], password, args['contact'], 'pembeli', args['email'], args['address'])
+        db.session.add(pembeliBaru)
+        db.session.commit()
+
+        pembeliBaru = marshal(pembeliBaru, Pembeli.response_pembeli)    
+
+        return {"status": "Created", "message": "Your account has been created", "input": pembeliBaru}, 201, {'Content-Type': 'application/json'}
 
     @jwt_required
     def get(self):
         pembeli = get_jwt_claims()
 
         qry = Pembeli.query.get(pembeli['id'])
+
+        if qry == None:
+            return {'status':'Not Found', 'message': 'DATA_NOT_FOUND'}, 404, {'Content-Type': 'application/json'}
+        
         result = marshal(qry, Pembeli.response_pembeli)
-        return result, 200, {'Content-Type': 'application/json'}
+        return {"status":"OK", "message":"Get Buyer Account", "buyer": result}, 200, {'Content-Type': 'application/json'}
             
     @jwt_required
-    def put(self):#, usernamePenbeli):
+    def put(self):
         pembeli = get_jwt_claims()
 
-        qry = Pembeli.query.get(pembeli['id'])  
+        qry = Pembeli.query.get(pembeli['id'])
+
+        if qry == None:
+            return {'status':'Not Found', 'message': 'DATA_NOT_FOUND'}, 404, {'Content-Type': 'application/json'}
+
         data_pembeli = marshal(qry, Pembeli.response_field)
 
-        # if usernamePembeli != pembeli['username']:
-        #     return {'message': "ACCESS_DENIED"}, 404, {'Content-Type': 'application/json'}
-
-        # else:
         parse =reqparse.RequestParser()
-        parse.add_argument('username', location='json', default = data_pembeli['username'])
-        parse.add_argument('password', location='json', default = data_pembeli['password'])
-        parse.add_argument('contact', location='json', default = data_pembeli['email'])
-        parse.add_argument('email', location='json', default = data_pembeli['email'])
-        parse.add_argument('address', location='json', default = data_pembeli['address'])
+        parse.add_argument('username', location='args', default = data_pembeli['username'])
+        parse.add_argument('password', location='args', default = data_pembeli['password'])
+        parse.add_argument('contact', location='args', default = data_pembeli['contact'])
+        parse.add_argument('email', location='args', default = data_pembeli['email'])
+        parse.add_argument('address', location='args', default = data_pembeli['address'])
 
         args = parse.parse_args()
 
@@ -63,17 +74,22 @@ class PembeliResource(Resource):
         qry.address = args['address']
         db.session.commit()
 
-        return {'message': 'DATA_UPDATED', 'input': marshal(qry, Pembeli.response_field)}, 200, {'Content-Type': 'application/json'}
+        return {'status':'Accepted', 'message': 'DATA_UPDATED', 'buyer': marshal(qry, Pembeli.response_pembeli)}, 202, {'Content-Type': 'application/json'}
 
     @jwt_required
-    def delete(self):#, usernamePenbeli):
+    def delete(self):
         pembeli = get_jwt_claims()
         qry = Pembeli.query.get(pembeli['id'])
 
+        if qry == None:
+            return {'status':'Not Found', 'message': 'DATA_NOT_FOUND'}, 404, {'Content-Type': 'application/json'}
+
+        data_pembeli = marshal(qry, Pembeli.response_pembeli)
+
         db.session.delete(qry)
         db.session.commit()
-        return {'message': 'DATA_DELETED'}, 200, {'Content-Type': 'application/json'}
+        return {'status':'OK', 'message': 'DATA_DELETED', 'buyer': data_pembeli}, 200, {'Content-Type': 'application/json'}
 
 
-api.add_resource(PembeliResource, '/pembeli/profile')
+api.add_resource(PembeliResource, '/pembeli')
